feat: shared NumberKeypad component, tap reps value to open keypad
Extract the expanded keypad UI from WeightKeypad into a reusable
NumberKeypad with configurable label and optional unit. Wire it into
RepsStepper so tapping the rep count opens the same keypad as weight
entry, avoiding many +/- taps for higher rep counts.
